+++ spec/support/jasmine.json
@@ -4,7 +4,7 @@
         "**/*[sS]pec.js"
     ],
     "helpers": [
-        "helpers/**/*.js"
+        "jasmine_helpers/**/*.js"
     ],
     "stopSpecOnExpectationFailure": false,
     "random": false
